+++ src/modules/webflow-membership.js
@@ -50,6 +50,26 @@ export class WfuMembership {
         return window.getCookie("wf_loggedin") || false;
     }
 
+    // Expanded login button
+    // Used on a containing DIV to expand the trigger area of 
+    // Webflow's Log-In / Log-Out button  
+    expandLoginButton = function($elem) {
+
+        // Get Webflow Login/Logout button
+        const $wfLoginButton = $elem.find("[data-wf-user-logout]");
+
+        // Setup click event handler on outer DIV
+        $elem.click(function() {
+            // Click inner element
+            $wfLoginButton.trigger('click');
+        });
+
+        // Also interecept and stop propogation so event is not doubled
+        $wfLoginButton.click(function(e) {
+            e.stopPropagation();
+        });
+
+    }
 
 }