+++ src/nocode/webflow-membership.js
@@ -11,13 +11,23 @@
 
 import { WfuMembership } from '../modules/webflow-membership.js';
 
+
+
 $(function () {
 
     console.debug("running WFU Membership");
 
     const wfuMembership = new WfuMembership();
+
+    // P
   
-    console.debug(`isLoggedIn = ${wfuMembership.isLoggedIn()}`); 
+    console.debug(`isLoggedIn = %c${wfuMembership.isLoggedIn()}`, "color: #ff0000;"); 
+
+
+
+
+
+    // Conditional visiblity
 
     // Show all elements tagged for logged-in users only
     $("[wfu-show-logged-in]").each(function() {
@@ -28,21 +38,23 @@ $(function () {
     // Show all elements tagged for logged-out users only
     $("[wfu-hide-logged-in]").each(function() { 
         if (!wfuMembership.isLoggedIn())
-        $(this).removeAttr("wfu-hide-logged-in"); 
+            $(this).removeAttr("wfu-hide-logged-in"); 
     });
 
+    // Expand login buttons
+    $("[wfu-login-button]").each(function() { 
+        
+        wfuMembership.expandLoginButton($(this));
 
-    // Expanded login button
-    // Used on a containing DIV to expand the trigger area of 
-    // Webflow's Log-In / Log-Out button  
-    $("[wfu-login-button]").click(function() {
+        /*
+        click(function() {
         $(this).find("[data-wf-user-logout]").trigger('click');
     })
     $("[wfu-login-button] [data-wf-user-logout]").click(function(e) {
         e.stopPropagation();
     });
-
-
+    */
+    });
 
 });